requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.5
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.4
Flask==2.1.2
Flask-Compress==1.12
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mysql-connector-python==8.0.29
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
platformdirs==2.5.2
plotly==5.8.0
protobuf==3.20.1
pylint==2.13.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tenacity==8.0.1
tomli==2.0.1
tweepy==4.9.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
